% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_lib.R
\name{ComputeIJCovarianceFromInfluence}
\alias{ComputeIJCovarianceFromInfluence}
\title{Return an estimate of the infinitesimal jackknife covariance estimate
of the frequentist variance of the posterior expectations of the parameters
in param_draws_mat.}
\usage{
ComputeIJCovarianceFromInfluence(infl_mat)
}
\arguments{
\item{infl_mat}{A matrix of n * cov(lp, par), with datapoints in rows and parameters
in columns.}
}
\value{
The IJ covariance matrix, which is an estimate of
N * Cov(E[params | x]), where N is the number of distinct exchangeable
observations.
}
\description{
Return an estimate of the infinitesimal jackknife covariance estimate
of the frequentist variance of the posterior expectations of the parameters
in param_draws_mat.
}
